requirements.txt
dbt-duckdb
duckdb
pandas
apache-airflow
pyspark
pyarrow